% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{find_in_gpl}
\alias{find_in_gpl}
\title{find gses by searching for match in whole table}
\usage{
find_in_gpl(x)
}
\arguments{
\item{x}{pattern to be used in stringr::str_detect}
}
\description{
find gses by searching for match in whole table
}

