% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scerden-geometadb.R
\docType{data}
\name{gds-subset}
\alias{gds-subset}
\title{scerden.geometadb package}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 772 rows and 6 columns.}
\source{
GEOmetadb bioconductor pkg
}
\usage{
gds-subset
}
\description{
subset of geometadb for Saccharomyces cerevisiae
}
\examples{
  `gds-susbet`
  gds
  gsm
  gse
  gpl
}
\keyword{datasets}

